🚀: Добавил роли Юзер, Автор, Админ и добавил проверки во вьюшках

+++ books_app/api/v1/views/author.py
@@ -1,22 +1,24 @@
 from rest_framework import status
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from api.v1.models.author import Author
+from api.v1.models.user import User
 from api.v1.serializers.author import AuthorsSerializer
 
 
 class AuthorsView(APIView):
-    permission_classes = [IsAuthenticatedOrReadOnly]
+    permission_classes = [IsAuthenticated]
 
     def get(self, request: Request):
-        authors = Author.objects.all()
+        authors = User.objects.filter(role=User.ROLES.AUTHOR.value)
         serializer = AuthorsSerializer(instance=authors, many=True, context={'request': request})
         return Response(serializer.data)
 
     def post(self, request: Request):
+        if not request.user.is_admin:
+            return Response({"detail": "Only admins can create authors"}, status=status.HTTP_403_FORBIDDEN)
         serializer = AuthorsSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -24,29 +26,35 @@ class AuthorsView(APIView):
 
 
 class AuthorView(APIView):
-    permission_classes = [IsAuthenticatedOrReadOnly]
+    permission_classes = [IsAuthenticated]
 
     def get(self, request: Request, author_id: int):
         try:
-            book = Author.objects.get(id=author_id)
-        except Author.DoesNotExist:
+            book = User.objects.get(id=author_id)
+        except User.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
         serializer = AuthorsSerializer(book, context={'request': request})
         return Response(serializer.data)
 
     def delete(self, request: Request, author_id: int):
+        if not request.user.is_admin:
+            return Response({"detail": "Only admins can delete another accounts"}, status=status.HTTP_403_FORBIDDEN)
         try:
-            author = Author.objects.get(id=author_id)
-        except Author.DoesNotExist:
+            author = User.objects.get(id=author_id)
+        except User.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
+        self.check_object_permissions(self.request, author)
         author.delete()
         return Response(status=status.HTTP_200_OK)
 
     def patch(self, request: Request, author_id):
+        if not (request.user.is_admin or request.user.id == author_id):
+            return Response({"detail": "Only admins can change another accounts"}, status=status.HTTP_403_FORBIDDEN)
         try:
-            author = Author.objects.get(id=author_id)
-        except Author.DoesNotExist:
+            author = User.objects.get(id=author_id)
+        except User.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
+        self.check_object_permissions(self.request, author)
         serializer = AuthorsSerializer(data=request.data, instance=author)
         serializer.is_valid(raise_exception=True)
         serializer.save()
